% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addMeta}
\alias{addMeta}
\title{Add the cell information into meta slot}
\usage{
addMeta(object, meta, meta.name = NULL)
}
\arguments{
\item{object}{CellChat object}

\item{meta}{cell information to be added}

\item{meta.name}{the name of column to be assigned}
}
\value{

}
\description{
Add the cell information into meta slot
}
