% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellChatDB.human}
\alias{CellChatDB.human}
\title{Ligand-receptor interactions in CellChat database for human}
\format{A list includes the ligand-receptor interactions}
\source{
\url{https://github.com/sqjin/CellChat/}
}
\usage{
CellChatDB.human
}
\description{
The ligand-receptor interaction database curated in CellChat tool
}
\keyword{datasets}
