% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellChatDB.zebrafish}
\alias{CellChatDB.zebrafish}
\title{Ligand-receptor interactions in CellChat database for Zebrafish}
\format{
A list includes the ligand-receptor interactions
}
\source{
\url{https://github.com/sqjin/CellChat/}
}
\usage{
CellChatDB.zebrafish
}
\description{
The ligand-receptor interaction database curated in CellChat tool
}
\keyword{datasets}
