% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{checkGeneSymbol}
\alias{checkGeneSymbol}
\title{check the official Gene Symbol}
\usage{
checkGeneSymbol(geneSet, geneIfo)
}
\arguments{
\item{geneSet}{gene set to check}

\item{geneIfo}{official Gene Symbol}
}
\description{
check the official Gene Symbol
}
