% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{computeCentralityLocal}
\alias{computeCentralityLocal}
\title{Compute Centrality measures for a signaling network}
\usage{
computeCentralityLocal(net)
}
\arguments{
\item{net}{compute the centrality measures on a specific signaling network given by a 2 or 3 dimemsional array net}
}
\value{

}
\description{
Compute Centrality measures for a signaling network
}
