% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeCommunProb}
\alias{computeCommunProb}
\title{Compute the communication probability/strength of any interacting cell groups}
\usage{
computeCommunProb(
  object,
  LR.use = NULL,
  raw.use = FALSE,
  population.size = TRUE,
  trim = NULL,
  nboot = 100,
  seed.use = 1L,
  Kh = 0.5
)
}
\arguments{
\item{object}{CellChat object}

\item{LR.use}{ligand-receptor interactions used in inferring communication network}

\item{raw.use}{whether use the raw data or the projected data}

\item{population.size}{whether consider the proportion of cells in each group across all sequenced cells. Set population.size = FALSE if analyzing sorting-enriched single cells, to remove the
potential artifact of population size. Set population.size = TRUE if analyzing unsorted single-cell transcriptomes. We reason that abundant cell populations tend to
send collectively stronger signals than the rare cell populations.}

\item{trim}{the fraction (0 to 0.25) of observations to be trimmed from each end of x before the mean is computed; default using the robust Tukey's trimean}

\item{nboot}{threshold of p-values}

\item{seed.use}{set a random seed. By default, sets the seed to 1.}

\item{Kh}{parameter in Hill function}
}
\value{

}
\description{
Compute the communication probability/strength of any interacting cell groups
}
