% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{computeEigengap}
\alias{computeEigengap}
\title{Compute the eigengap of a given matrix for inferring the number of clusters}
\usage{
computeEigengap(CM, tau = NULL, tol = 0.01)
}
\arguments{
\item{CM}{consensus matrix}

\item{tau}{truncated consensus matrix}

\item{tol}{tolerance}
}
\value{

}
\description{
Compute the eigengap of a given matrix for inferring the number of clusters
}
