% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeExpr_agonist}
\alias{computeExpr_agonist}
\title{Modeling the effect of agonist on the ligand-receptor interaction}
\usage{
computeExpr_agonist(
  data.use,
  pairLRsig,
  cofactor_input,
  group,
  index.agonist,
  Kh,
  FunMean
)
}
\arguments{
\item{data.use}{data matrix}

\item{pairLRsig}{the L-R interactions}

\item{cofactor_input}{the cofactor_input from CellChatDB}

\item{group}{a factor defining the cell groups}

\item{index.agonist}{the index of agonist in the database}

\item{Kh}{a parameter in Hill function}

\item{FunMean}{the function for computing mean expression per group}
}
\value{

}
\description{
Modeling the effect of agonist on the ligand-receptor interaction
}
