% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeExpr_coreceptor}
\alias{computeExpr_coreceptor}
\title{Modeling the effect of coreceptor on the ligand-receptor interaction}
\usage{
computeExpr_coreceptor(
  cofactor_input,
  data.use,
  coreceptor.all,
  index.coreceptor
)
}
\arguments{
\item{cofactor_input}{the cofactor_input from CellChatDB}

\item{data.use}{data matrix}

\item{coreceptor.all}{all the coreceptor in the database}

\item{index.coreceptor}{the index of coreceptor in the database}
}
\value{

}
\description{
Modeling the effect of coreceptor on the ligand-receptor interaction
}
