% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeExpr_LR}
\alias{computeExpr_LR}
\title{Compute the expression of ligands or receptors using geometric mean}
\usage{
computeExpr_LR(geneLR, data.use, complex_input)
}
\arguments{
\item{geneLR}{a char vector giving a set of ligands or receptors}

\item{data.use}{data matrix (row are genes and columns are cells or cell groups)}

\item{complex_input}{the complex_input from CellChatDB}
}
\value{

}
\description{
Compute the expression of ligands or receptors using geometric mean
}
