% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{computeLaplacian}
\alias{computeLaplacian}
\title{Compute eigenvalues of associated Laplacian matrix of a given matrix}
\usage{
computeLaplacian(CM, tol = 0.01)
}
\arguments{
\item{CM}{consensus matrix}

\item{tol}{tolerance}
}
\value{

}
\description{
Compute eigenvalues of associated Laplacian matrix of a given matrix
}
