% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{computeNetD_structure}
\alias{computeNetD_structure}
\title{Compute the structural distance between two signaling networks}
\usage{
computeNetD_structure(g, h, w1 = 0.45, w2 = 0.45, w3 = 0.1)
}
\arguments{
\item{g}{a graph object of one signaling network}

\item{h}{a graph object of another signaling network}

\item{w1}{parameter}

\item{w2}{parameter}

\item{w3}{parameter}
}
\value{

}
\description{
Compute the structural distance between two signaling networks
}
