% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{computeNetSimilarity}
\alias{computeNetSimilarity}
\title{Compute signaling network similarity for any pair of signaling networks}
\usage{
computeNetSimilarity(
  object,
  slot.name = "netP",
  type = c("functional", "structural"),
  k = NULL,
  thresh = NULL
)
}
\arguments{
\item{object}{CellChat object}

\item{slot.name}{the slot name of object that is used to compute centrality measures of signaling networks}

\item{type}{"functional","structural"}

\item{k}{the number of nearest neighbors}

\item{thresh}{the fraction (0 to 0.25) of interactions to be trimmed before computing network similarity}
}
\value{

}
\description{
Compute signaling network similarity for any pair of signaling networks
}
