% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{computeNetSimilarityPairwise}
\alias{computeNetSimilarityPairwise}
\title{Compute signaling network similarity for a pair of datasets}
\usage{
computeNetSimilarityPairwise(
  object,
  slot.name = "netP",
  type = c("functional", "structural"),
  k = NULL,
  thresh = 0.25
)
}
\arguments{
\item{object}{A merged CellChat object}

\item{slot.name}{the slot name of object that is used to compute centrality measures of signaling networks}

\item{type}{"functional","structural"}

\item{k}{the number of nearest neighbors}

\item{thresh}{a range of the number of patterns}
}
\value{

}
\description{
Compute signaling network similarity for a pair of datasets
}
