% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{extractGene}
\alias{extractGene}
\title{Extract the genes involved in CellChatDB}
\usage{
extractGene(CellChatDB)
}
\arguments{
\item{CellChatDB}{CellChatDB databse used in the analysis}
}
\value{

}
\description{
Extract the genes involved in CellChatDB
}
