% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{extractGeneSubset}
\alias{extractGeneSubset}
\title{Extract the gene name}
\usage{
extractGeneSubset(geneSet, complex_input, geneIfo)
}
\arguments{
\item{geneSet}{gene set}

\item{complex_input}{complex in CellChatDB databse}

\item{geneIfo}{official gene symbol}
}
\value{

}
\description{
Extract the gene name
}
