% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{extractGeneSubsetFromPair}
\alias{extractGeneSubsetFromPair}
\title{Extract the signaling gene names from ligand-receptor pairs}
\usage{
extractGeneSubsetFromPair(
  pairLR,
  object = NULL,
  complex_input = NULL,
  geneInfo = NULL,
  combined = TRUE
)
}
\arguments{
\item{pairLR}{data frame must contain columns named `ligand` and `receptor`}

\item{object}{a CellChat object}

\item{complex_input}{complex in CellChatDB databse}

\item{geneInfo}{official gene symbol}

\item{combined}{whether combining the ligand genes and receptor genes}
}
\value{

}
\description{
Extract the signaling gene names from ligand-receptor pairs
}
