% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{extractLRfromGenes}
\alias{extractLRfromGenes}
\title{Extract L-R pairs associated with a given gene set}
\usage{
extractLRfromGenes(geneSet, db)
}
\arguments{
\item{geneSet}{a vector of genes}

\item{db}{one of the CellChatDB databases (e.g., CellChatDB.human, CellChatDB.mouse...)}
}
\description{
Extract L-R pairs associated with a given gene set
}
