% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{geometricMean}
\alias{geometricMean}
\title{Compute the geometric mean}
\usage{
geometricMean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{whether remove na}
}
\value{

}
\description{
Compute the geometric mean
}
