% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{identifyOverExpressedInteractions}
\alias{identifyOverExpressedInteractions}
\title{Identify over-expressed ligand-receptor interactions}
\usage{
identifyOverExpressedInteractions(object, features = NULL)
}
\arguments{
\item{object}{CellChat object}

\item{features}{a vector of features}
}
\value{

}
\description{
Identify over-expressed ligand-receptor interactions
}
