% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{l1NormalizeColumns}
\alias{l1NormalizeColumns}
\title{Column-normalize a sparse, symmetric matrix (using the l1 norm) so that each
column sums to 1.}
\usage{
l1NormalizeColumns(A)
}
\arguments{
\item{A}{matrix}
}
\value{
column-normalized sparse matrix object
}
\description{
Column-normalize a sparse, symmetric matrix (using the l1 norm) so that each
column sums to 1.
}
\keyword{internal}
