% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{l1NormalizeRows}
\alias{l1NormalizeRows}
\title{Row-normalize a sparse, symmetric matrix (using the l1 norm) so that each
row sums to 1.}
\usage{
l1NormalizeRows(A)
}
\arguments{
\item{A}{matrix}
}
\value{
row-normalized sparse matrix object
}
\description{
Row-normalize a sparse, symmetric matrix (using the l1 norm) so that each
row sums to 1.
}
\keyword{internal}
