% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChat_class.R
\name{mergeCellChat}
\alias{mergeCellChat}
\title{Merge CellChat objects}
\usage{
mergeCellChat(object1, object2, add.names = NULL)
}
\arguments{
\item{object1}{CellChat object of dataset 1}

\item{object2}{CellChat object of dataset 2}

\item{add.names}{the names of each dataset}
}
\value{

}
\description{
Merge CellChat objects
}
