% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{mergeInteractions}
\alias{mergeInteractions}
\title{Compute the number of interactions/interaction strength between cell types based on their associated cell subpopulations}
\usage{
mergeInteractions(object, group.merged)
}
\arguments{
\item{object}{CellChat object}

\item{group.merged}{a factor defining the group for merging different clusters/subpopulations}
}
\value{
An updated slot `net` by adding three elements:

`count.merged`: the number of interactions between cell types (i.e., merged cell groups)

`weight.merged`: interaction strength between cell types (i.e., merged cell groups)

`group.merged` the defined group for merging different clusters/subpopulations
}
\description{
Compute the number of interactions/interaction strength between cell types based on their associated cell subpopulations
}
