% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{modify_vlnplot}
\alias{modify_vlnplot}
\title{modified vlnplot}
\usage{
modify_vlnplot(
  object,
  features,
  idents = NULL,
  split.by = NULL,
  cols = NULL,
  show.text.y = TRUE,
  line.size = NULL,
  pt.size = 0,
  plot.margin = margin(0, 0, 0, 0, "cm"),
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{features}{Features to plot (gene expression, metrics)}

\item{idents}{Which classes to include in the plot (default is all)}

\item{split.by}{Name of a metadata column to split plot by;}

\item{cols}{defining the color for each cell group}

\item{show.text.y}{whther show y-axis text}

\item{line.size}{line width in the violin plot}

\item{pt.size}{size of the dots}

\item{plot.margin}{adjust the white space between each plot}

\item{...}{pass any arguments to VlnPlot in Seurat}
}
\description{
modified vlnplot
}
