% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{mycircle}
\alias{mycircle}
\title{generate circle symbol}
\usage{
mycircle(coords, v = NULL, params)
}
\arguments{
\item{coords}{coordinates of points}

\item{v}{vetex}

\item{params}{parameters}
}
\value{

}
\description{
generate circle symbol
}
