% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{netAnalysis_signalingRole}
\alias{netAnalysis_signalingRole}
\title{Identification of dominant senders, receivers, mediators and influencers in the intercellular communication network}
\usage{
netAnalysis_signalingRole(
  object,
  slot.name = "netP",
  net = NULL,
  net.name = NULL
)
}
\arguments{
\item{object}{CellChat object}

\item{slot.name}{the slot name of object that is used to compute centrality measures of signaling networks}

\item{net}{compute the centrality measures on a specific signaling network given by a 2 or 3 dimemsional array net}

\item{net.name}{a character vector giving the name of signaling networks}
}
\value{

}
\description{
Identification of dominant senders, receivers, mediators and influencers in the intercellular communication network
}
