% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual}
\alias{netVisual}
\title{Visualize the inferred cell-cell communication network}
\usage{
netVisual(
  object,
  signaling,
  signaling.name = NULL,
  vertex.receiver = NULL,
  top = 1,
  color.use = NULL,
  from = NULL,
  to = NULL,
  bidirection = FALSE,
  remove.isolate = FALSE,
  vertex.size = 20,
  layout = c("hierarchy", "circle"),
  height = 5,
  thresh = 0.05,
  pt.title = 12,
  title.space = 6,
  vertex.label.cex = 0.8,
  out.format = c("svg", "png")
)
}
\arguments{
\item{object}{CellChat object}

\item{signaling}{a signaling pathway name}

\item{signaling.name}{alternative signaling pathway name to show on the plot}

\item{vertex.receiver}{a numeric vector giving the index of the cell groups as targets in the first hierarchy plot}

\item{top}{the fraction of interactions to show (0 < top <= 1)}

\item{color.use}{the character vector defining the color of each cell group}

\item{from}{a vector giving the index or the name of source cell groups when using circle plot}

\item{to}{a vector giving the index or the name of target cell groups.}

\item{bidirection}{whether show the bidirectional communication, i.e., both 'from'->'to' and 'to'->'from'.}

\item{remove.isolate}{whether remove the isolate nodes in the communication network}

\item{vertex.size}{The size of vertex}

\item{layout}{"hierarchy" or "circle"}

\item{height}{height of plot}

\item{thresh}{threshold of the p-value for determining significant interaction}

\item{pt.title}{font size of the text}

\item{title.space}{the space between the title and plot}

\item{vertex.label.cex}{The label size of vertex in the network}

\item{out.format}{the format of output figures: svg, png and pdf}
}
\value{

}
\description{
Automatically save plots in the current working directory
}
\examples{

}
