% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual_individual}
\alias{netVisual_individual}
\title{Visualize the inferred signaling network of individual L-R pairs}
\usage{
netVisual_individual(
  object,
  signaling,
  signaling.name = NULL,
  vertex.receiver = NULL,
  color.use = NULL,
  vertex.size = 20,
  layout = c("hierarchy", "circle"),
  height = 5,
  thresh = 0.05
)
}
\arguments{
\item{object}{CellChat object}

\item{signaling}{a signaling pathway name}

\item{signaling.name}{alternative signaling pathway name to show on the plot}

\item{vertex.receiver}{a numeric vector giving the index of the cell groups as targets in the first hierarchy plot}

\item{color.use}{the character vector defining the color of each cell group}

\item{vertex.size}{The size of vertex}

\item{layout}{"hierarchy" or "circle"}

\item{height}{height of plot}

\item{thresh}{threshold of the p-value for determining significant interaction}
}
\value{

}
\description{
Visualize the inferred signaling network of individual L-R pairs
}
\examples{

}
