% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{node_distance}
\alias{node_distance}
\title{compute the node distance matrix}
\usage{
node_distance(g)
}
\arguments{
\item{g}{a graph objecct}
}
\value{

}
\description{
compute the node distance matrix
}
