% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pieChart}
\alias{pieChart}
\title{Plot pie chart}
\usage{
pieChart(df, label.size = 2.5, color.use = NULL, title = "")
}
\arguments{
\item{df}{a dataframe}

\item{label.size}{a character}

\item{color.use}{the name of the variable in CellChatDB interaction_input}

\item{title}{the title of plot}
}
\value{

}
\description{
Plot pie chart
}
