% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PPI.human}
\alias{PPI.human}
\title{Human Protein-Protein interactions}
\format{
A square matrix where A_{ij}=1 if protein i interacts with protein j
}
\source{
\url{http://www.string-db.org/}
}
\usage{
PPI.human
}
\description{
An adjacency matrix of the high confidence known experimental interactions
between human proteins on STRINGdb.
}
\keyword{datasets}
