% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{preProcMultiomics}
\alias{preProcMultiomics}
\title{Preprocessing multi-omics data and preparing the L-R database}
\usage{
preProcMultiomics(data.list, db, do.sparse = TRUE)
}
\arguments{
\item{data.list}{a list consisting of multi-omics data (e.g., RNA & ADT)}

\item{db}{one of the CellChatDB databases: CellChatDB.human, CellChatDB.mouse, CellChatDB.zebrafish}

\item{do.sparse}{whether to use sparse format}
}
\description{
Preprocessing multi-omics data and preparing the L-R database
}
