% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{projectData}
\alias{projectData}
\title{Project gene expression data onto a protein-protein interaction network}
\usage{
projectData(
  object,
  adjMatrix,
  alpha = 0.5,
  normalizeAdjMatrix = c("rows", "columns")
)
}
\arguments{
\item{object}{CellChat object}

\item{adjMatrix}{adjacency matrix of protein-protein interaction network to use}

\item{alpha}{numeric in [0,1]}

\item{normalizeAdjMatrix}{how to normalize the adjacency matrix
possible values are 'rows' (in-degree)
and 'columns' (out-degree)}
}
\value{
projected gene expression matrix
}
\description{
Project gene expression data onto a protein-protein interaction network
}
