% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{projectFromNetworkRecombine}
\alias{projectFromNetworkRecombine}
\title{Combine gene expression from projected space (that of the network) with the
expression of genes that were not projected (not present in network)}
\usage{
projectFromNetworkRecombine(original_expression, projected_expression)
}
\arguments{
\item{original_expression}{the non-projected expression}

\item{projected_expression}{the projected gene expression, in the space
of the genes defined by the network}
}
\value{
a matrix in the dimensions of original_expression, where values that
        are present in projected_expression are copied from there.
}
\description{
Combine gene expression from projected space (that of the network) with the
expression of genes that were not projected (not present in network)
}
\keyword{internal}
