% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{projectOnNetwork}
\alias{projectOnNetwork}
\title{Project the gene expression matrix onto a lower space
of the genes defined in the projecting network}
\usage{
projectOnNetwork(gene_expression, new_features, missing.value = 0)
}
\arguments{
\item{gene_expression}{gene expression matrix}

\item{new_features}{the genes in the network, on which to project
the gene expression matrix}

\item{missing.value}{value to assign to genes that are in network,
but missing from gene expression matrix}
}
\value{
the gene expression matrix projected onto the gene space defined by new_features
}
\description{
Project the gene expression matrix onto a lower space
of the genes defined in the projecting network
}
\keyword{internal}
