% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{randomWalkBySolve}
\alias{randomWalkBySolve}
\title{project data on graph by solving the linear equation (I - alpha*A) * E_sm = E * (1-alpha)}
\usage{
randomWalkBySolve(E, A, alpha, normalizeAjdMatrix = c("rows", "columns"))
}
\arguments{
\item{E}{initial data matrix [NxM]}

\item{A}{adjacency matrix of graph to network project on will be column-normalized.}

\item{alpha}{projecting coefficient (1 - restart probability of random walk)}
}
\value{
network-projected gene expression
}
\description{
project data on graph by solving the linear equation (I - alpha*A) * E_sm = E * (1-alpha)
}
\keyword{internal}
