% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rankNetPairwise}
\alias{rankNetPairwise}
\title{Rank ligand-receptor interactions for any pair of two cell groups}
\usage{
rankNetPairwise(object, LR.use = NULL)
}
\arguments{
\item{object}{CellChat object}

\item{LR.use}{ligand-receptor interactions used in inferring communication network}
}
\value{

}
\description{
Rank ligand-receptor interactions for any pair of two cell groups
}
