% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rankSimilarity}
\alias{rankSimilarity}
\title{Rank the similarity of the shared signaling pathways based on their joint manifold learning}
\usage{
rankSimilarity(
  object,
  slot.name = "netP",
  type = c("functional", "structural"),
  comparison = c(1, 2),
  x.rotation = 90,
  title = NULL,
  color.use = NULL,
  bar.w = NULL,
  font.size = 8
)
}
\arguments{
\item{object}{CellChat object}

\item{slot.name}{the slot name of object that is used to compute centrality measures of signaling networks}

\item{type}{"functional","structural"}

\item{comparison}{a numerical vector giving the datasets for comparison}

\item{x.rotation}{rotation of x-labels}

\item{title}{main title of the plot}

\item{color.use}{defining the color for each cell group}

\item{bar.w}{the width of bar plot}

\item{font.size}{font size}
}
\value{

}
\description{
Rank the similarity of the shared signaling pathways based on their joint manifold learning
}
