% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runCellChatApp}
\alias{runCellChatApp}
\title{Generate a Shiny App for interactive exploration of CellChat's outputs}
\usage{
runCellChatApp(object, ...)
}
\arguments{
\item{object}{CellChat object}

\item{...}{Other parameters of `shinyApp` function from shiny R package}
}
\value{
A Shiny app object on the basis of one CellChat object
}
\description{
Generate a Shiny App for interactive exploration of CellChat's outputs
}
