% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scaleData}
\alias{scaleData}
\title{Scale the data}
\usage{
scaleData(data.use, do.center = T)
}
\arguments{
\item{data.use}{input data}

\item{do.center}{whether center the values}
}
\description{
Scale the data
}
