% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scaleMat}
\alias{scaleMat}
\title{Scale a data matrix}
\usage{
scaleMat(x, scale, na.rm = TRUE)
}
\arguments{
\item{x}{data matrix}

\item{scale}{the method to scale the data}

\item{na.rm}{whether remove na}
}
\value{

}
\description{
Scale a data matrix
}
