% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scPalette}
\alias{scPalette}
\title{Generate colors from a customed color palette}
\usage{
scPalette(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
A color palette for plotting
}
\description{
Generate colors from a customed color palette
}
