% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{searchPair}
\alias{searchPair}
\title{Subset the ligand-receptor interactions for given specific signals in CellChatDB}
\usage{
searchPair(
  signaling = c(),
  pairLR.use,
  key = c("pathway_name", "ligand"),
  matching.exact = FALSE,
  pair.only = TRUE
)
}
\arguments{
\item{signaling}{a character vector}

\item{pairLR.use}{a dataframe containing ligand-receptor interactions}

\item{key}{the keyword to match}

\item{matching.exact}{whether perform exact matching}

\item{pair.only}{whether only return ligand-receptor pairs without cofactors}
}
\value{

}
\description{
Subset the ligand-receptor interactions for given specific signals in CellChatDB
}
