% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{selectK}
\alias{selectK}
\title{Select the number of the patterns}
\usage{
selectK(
  object,
  slot.name = "netP",
  pattern = c("outgoing", "incoming"),
  title.name = NULL,
  do.facet = TRUE,
  k.range = seq(2, 10),
  nrun = 30,
  seed.use = 10
)
}
\arguments{
\item{object}{CellChat object}

\item{slot.name}{the slot name of object that is used to compute centrality measures of signaling networks}

\item{pattern}{"outgoing" or "incoming"}

\item{title.name}{title of plot}

\item{do.facet}{whether use facet plot showing the two measures}

\item{k.range}{a range of the number of patterns}

\item{nrun}{number of runs when performing NMF}

\item{seed.use}{seed when performing NMF}
}
\value{

}
\description{
Select the number of the patterns
}
