% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setIdent}
\alias{setIdent}
\title{Set the default identity of cells}
\usage{
setIdent(object, ident.use = NULL, levels = NULL, display.warning = TRUE)
}
\arguments{
\item{object}{CellChat object}

\item{ident.use}{the name of the variable in object.meta;}

\item{levels}{set the levels of factor}

\item{display.warning}{whether display the warning message}
}
\value{

}
\description{
Set the default identity of cells
}
