% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChat_class.R
\docType{methods}
\name{show,CellChat-method}
\alias{show,CellChat-method}
\title{show method for CellChat}
\usage{
\S4method{show}{CellChat}(object)
}
\arguments{
\item{object}{object}

\item{CellChat}{object}

\item{show}{show the object}
}
\description{
show method for CellChat
}
