% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{showDatabaseCategory}
\alias{showDatabaseCategory}
\title{Show the description of CellChatDB databse}
\usage{
showDatabaseCategory(CellChatDB, nrow = 1)
}
\arguments{
\item{CellChatDB}{CellChatDB databse}

\item{nrow}{the number of rows in the plot}
}
\value{

}
\description{
Show the description of CellChatDB databse
}
