% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sketchData}
\alias{sketchData}
\title{Downsampling single cell data}
\usage{
sketchData(X, percent, dimPC = 50)
}
\arguments{
\item{X}{input data (samples in rows, features in columns)}

\item{percent}{the percent of data to sketch}

\item{dimPC}{the number of components to use}
}
\value{

}
\description{
Downsampling single cell data
}
