% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{subsetData}
\alias{subsetData}
\title{Subset the expression data of signaling genes in CellChatDB}
\usage{
subsetData(object, features = NULL)
}
\arguments{
\item{object}{CellChat object}

\item{features}{default = NULL: subset the expression data of signaling genes in CellChatDB}
}
\value{

}
\description{
Subset the expression data of signaling genes in CellChatDB
}
