% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{subsetDB}
\alias{subsetDB}
\title{Subset CellChatDB databse by only including interactions of interest}
\usage{
subsetDB(CellChatDB, search = c(), key = "annotation", non_protein = FALSE)
}
\arguments{
\item{CellChatDB}{CellChatDB databse}

\item{search}{a character vector; When `key` is a vector, the `search` should be a list with the size being `length(key)`, where each element is a character vector}

\item{key}{a character vector and each element should be one of the column names of the interaction_input from CellChatDB.}

\item{non_protein}{whether to use the non-protein signaling for CellChat analysis. By default, non_protein = FALSE because most of non-protein signaling are the special synaptic signaling interactions that can only be used when inferring neuron-neuron communication.}
}
\description{
Subset CellChatDB databse by only including interactions of interest
}
