% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{thresholdedMean}
\alias{thresholdedMean}
\title{Compute the average expression per cell group when the percent of expressing cells per cell group larger than a threshold}
\usage{
thresholdedMean(x, trim = 0.1, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{trim}{the percent of expressing cells per cell group to be considered as zero}

\item{na.rm}{whether remove na}
}
\value{

}
\description{
Compute the average expression per cell group when the percent of expressing cells per cell group larger than a threshold
}
