% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{triMean}
\alias{triMean}
\title{Compute the Tukey's trimean}
\usage{
triMean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{whether remove na}
}
\value{

}
\description{
Compute the Tukey's trimean
}
