% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{updateCCC_score}
\alias{updateCCC_score}
\title{Update the cell-cell communication array from a customized cell-cell-communication scores between different cell groups}
\usage{
updateCCC_score(object, net)
}
\arguments{
\item{object}{CellChat object}

\item{net}{a data frame with at least five columns named as `source`,`target`,`ligand`,`receptor` and `score`, which defines the customized cell-cell-communication scores between different cell groups.
a p-value column named `pval`, and additional columns named `interaction_name` and `interaction_name_2` can be also provided.}
}
\value{
a CellChat object with updated slot `net` and slot `DB` if db is not NULL.
}
\description{
Users may also check the `updateCellChatDB` function for integrating other resources or utilizing a custom database
}
