% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChat_class.R
\name{updateCellChat}
\alias{updateCellChat}
\title{Update a single CellChat object}
\usage{
updateCellChat(object)
}
\arguments{
\item{object}{CellChat object}
}
\value{
a updated CellChat object
}
\description{
Update: object@var.features is now object@var.features$features; object@net$sum is now object@net$weight if `aggregateNet` has been run.
}
