% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{updateClusterLabels}
\alias{updateClusterLabels}
\title{Update and re-order the cell group names after running `computeCommunProb`}
\usage{
updateClusterLabels(
  object,
  old.cluster.name = NULL,
  new.cluster.name = NULL,
  new.order = NULL,
  new.cluster.metaname = "new.labels"
)
}
\arguments{
\item{object}{CellChat object}

\item{old.cluster.name}{A vector defining old cell group labels in `object@idents`; Default = NULL, which will use `levels(object@idents)`}

\item{new.cluster.name}{A vector defining new cell group labels to rename}

\item{new.order}{reset order of cell group labels}

\item{new.cluster.metaname}{assign a name of the new labels, which will be the column name of new labels in `object@meta`}
}
\value{
An updated CellChat object
}
\description{
Update and re-order the cell group names after running `computeCommunProb`
}
