% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{aggregateNet}
\alias{aggregateNet}
\title{Calculate the aggregated network by counting the number of links or summarizing the communication probability}
\usage{
aggregateNet(object, thresh = 0.05)
}
\arguments{
\item{object}{CellChat object}

\item{thresh}{threshold of the p-value for determining significant interaction}
}
\value{

}
\description{
Calculate the aggregated network by counting the number of links or summarizing the communication probability
}
