% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChat_class.R
\docType{class}
\name{CellChat-class}
\alias{CellChat-class}
\alias{CellChat}
\title{The key slots used in the CellChat object are described below.}
\description{
The key slots used in the CellChat object are described below.
}
\section{Slots}{

\describe{
\item{\code{data.raw}}{raw data matrix, one per dataset (Genes should be in rows and cells in columns)}

\item{\code{data}}{normalized matrix (genes by cells)}

\item{\code{data.signaling}}{normalized matrix used in the CellChat analysis}

\item{\code{data.scale}}{scaled matrix}

\item{\code{data.project}}{projected data}

\item{\code{net}}{a three-dimensional array P (K×K×N), where K is the number of cell groups and N is the number of ligand-receptor pairs. Each row of P indicates the communication probability originating from the sender cell group to other cell groups.}

\item{\code{netP}}{communication networks on signaling pathway level}

\item{\code{DB}}{ligand-receptor interaction database used in the analysis}

\item{\code{LR}}{a list of information related with ligand-receptor pairs}

\item{\code{meta}}{data frame storing the information associated with each cell}

\item{\code{idents}}{a factor defining the cell identity}

\item{\code{var.features}}{informative features to be used}

\item{\code{dr}}{List of the reduced 2D coordinates, one per method, e.g., umap/tsne/dm}

\item{\code{options}}{List of parameters used throughout analysis}
}}

