% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeCommunProb}
\alias{computeCommunProb}
\title{Compute the communication probability/strength of any interacting cell groups}
\usage{
computeCommunProb(
  object,
  LR.use = NULL,
  trim = NULL,
  nboot = 100,
  seed.use = 1L,
  Kh = 0.5
)
}
\arguments{
\item{object}{CellChat object}

\item{LR.use}{ligand-receptor interactions used in inferring communication network}

\item{trim}{the fraction (0 to 0.25) of observations to be trimmed from each end of x before the mean is computed; default using the robust Tukey's trimean}

\item{nboot}{threshold of p-values}

\item{seed.use}{set a random seed. By default, sets the seed to 1.}

\item{Kh}{parameter in Hill function}
}
\value{

}
\description{
Compute the communication probability/strength of any interacting cell groups
}
