% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeCommunProbPathway}
\alias{computeCommunProbPathway}
\title{Compute the communication probability on signaling pathway level by summarizing all related ligands/receptors}
\usage{
computeCommunProbPathway(object, thresh = 0.05)
}
\arguments{
\item{object}{CellChat object}

\item{thresh}{threshold of the p-value for determining significant interaction}
}
\value{

}
\description{
Compute the communication probability on signaling pathway level by summarizing all related ligands/receptors
}
