% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeExpr_agonist}
\alias{computeExpr_agonist}
\title{Modeling the effect of agonist on the ligand-receptor interaction}
\usage{
computeExpr_agonist(data.use, pairLRsig, cofactor_input, index.agonist, Kh, n)
}
\arguments{
\item{data.use}{data matrix}

\item{pairLRsig}{the L-R interactions}

\item{cofactor_input}{the cofactor_input from CellChatDB}

\item{index.agonist}{the index of agonist in the database}

\item{Kh}{a parameter in Hill function}

\item{n}{Hill coefficient}
}
\value{

}
\description{
Modeling the effect of agonist on the ligand-receptor interaction
}
