% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeExpr_antagonist}
\alias{computeExpr_antagonist}
\title{Modeling the effect of antagonist on the ligand-receptor interaction}
\usage{
computeExpr_antagonist(
  data.use,
  pairLRsig,
  cofactor_input,
  index.antagonist,
  Kh,
  n
)
}
\arguments{
\item{data.use}{data matrix}

\item{pairLRsig}{the L-R interactions}

\item{cofactor_input}{the cofactor_input from CellChatDB}

\item{index.antagonist}{the index of antagonist in the database}

\item{Kh}{a parameter in Hill function}

\item{n}{Hill coefficient}
}
\value{

}
\description{
Modeling the effect of antagonist on the ligand-receptor interaction
}
