% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeExpr_complex}
\alias{computeExpr_complex}
\title{Compute the expression of complex using geometric mean}
\usage{
computeExpr_complex(complex_input, data.use, complex)
}
\arguments{
\item{complex_input}{the complex_input from CellChatDB}

\item{data.use}{data matrix}

\item{complex}{the names of complex}
}
\value{

}
\description{
Compute the expression of complex using geometric mean
}
