% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeExpr_coreceptor}
\alias{computeExpr_coreceptor}
\title{Modeling the effect of coreceptor on the ligand-receptor interaction}
\usage{
computeExpr_coreceptor(cofactor_input, data.use, pairLRsig, type = c("A", "I"))
}
\arguments{
\item{cofactor_input}{the cofactor_input from CellChatDB}

\item{data.use}{data matrix}

\item{pairLRsig}{a data frame giving ligand-receptor interactions}

\item{type}{when type == "A", computing expression of co-activation receptor; when type == "I", computing expression of co-inhibition receptor.}
}
\value{

}
\description{
Modeling the effect of coreceptor on the ligand-receptor interaction
}
