% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeRegionDistance}
\alias{computeRegionDistance}
\title{Compute the region distance based on the spatial locations of each splot/cell of the spatial transcriptomics}
\usage{
computeRegionDistance(
  coordinates,
  meta,
  trim = 0.1,
  interaction.range = NULL,
  ratio = NULL,
  tol = NULL,
  k.min = 10,
  contact.knn = TRUE,
  contact.knn.k = 6
)
}
\arguments{
\item{coordinates}{a data matrix in which each row gives the spatial locations/coordinates of each cell/spot}

\item{meta}{a data frame including at least two columns named `group` and `slices`. `meta$group` is a factor vector defining the regions/labels of each cell/spot. `meta$slices` is a factor vector defining the slice labels of each dataset.}

\item{trim}{the fraction (0 to 0.25) of observations to be trimmed from each end of x before computing the average distance per cell group.}

\item{interaction.range}{The maximum interaction/diffusion range of ligands. This hard threshold is used to filter out the connections between spatially distant regions}

\item{ratio}{a numerical vector giving the conversion factor when converting spatial coordinates from Pixels or other units to Micrometers (i.e.,Microns).

For example, setting `ratio = 0.18` indicates that 1 pixel equals 0.18um in the coordinates.
For 10X visium, it is the ratio of the theoretical spot size (i.e., 65um) over the number of pixels that span the diameter of a theoretical spot size in the full-resolution image (i.e., 'spot.size.fullres' in the 'scalefactors_json.json' file).}

\item{tol}{a numerical vector giving the tolerance factor to increase the robustness when comparing the center-to-center distance against the `interaction.range`. This can be the half value of cell/spot size in the unit of um.

For example, for 10X visium, `tol` can be set as `65/2`; for slide-seq, `tol` can be set as `10/2`.
If the cell/spot size is not known, we provide a function `computeCellDistance` to compute the center-to-center distance. `tol` can be the the half value of the minimum center-to-center distance.}

\item{k.min}{the minimum number of interacting cell pairs required for defining adjacent cell groups}

\item{contact.knn}{whether determining spatially proximal cell regions based on the k-nearest neighbors (knn). Users can set `contact.knn = TRUE` if inferring contact-dependent signaling or preferring interactions within spatial neighbors.}

\item{contact.knn.k}{Number of neighbors in a knn graph, which is used to restrict the contact-dependent signaling within the neatest neighbors. For 10X visium, contact.knn.k = 6.}
}
\value{
A list including a square matrix giving the pairwise region distance and an adjacent matrix obtained from the k.spatial neatest neighbors
}
\description{
Compute the region distance based on the spatial locations of each splot/cell of the spatial transcriptomics
}
