% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{computeRegionDistance}
\alias{computeRegionDistance}
\title{Compute the region distance based on the spatial locations of each splot/cell of the spatial transcriptomics}
\usage{
computeRegionDistance(
  coordinates,
  group,
  trim = 0.1,
  interaction.length = NULL,
  spot.size = NULL,
  spot.size.fullres = NULL,
  k.min = 10
)
}
\arguments{
\item{coordinates}{a data matrix in which each row gives the spatial locations/coordinates of each cell/spot}

\item{group}{a factor vector defining the regions/labels of each cell/spot}

\item{trim}{the fraction (0 to 0.25) of observations to be trimmed from each end of x before computing the average distance per cell group.}

\item{interaction.length}{The maximum interaction/diffusion length of ligands. This hard threshold is used to filter out the connections between spatially distant regions}

\item{spot.size}{theoretical spot size; e.g., 10x Visium (spot.size = 65 microns)}

\item{spot.size.fullres}{The number of pixels that span the diameter of a theoretical spot size in the original,full-resolution image.}

\item{k.min}{the minimum number of interacting cell pairs required for defining adjacent cell groups}
}
\value{
A square matrix giving the pairwise region distance
}
\description{
Compute the region distance based on the spatial locations of each splot/cell of the spatial transcriptomics
}
