% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChat_class.R
\name{createCellChat}
\alias{createCellChat}
\title{creat a new CellChat object}
\usage{
createCellChat(data, do.sparse = T)
}
\arguments{
\item{data}{raw data matrix, a single-cell transcriptomic data}

\item{do.sparse}{whether use sparse format}
}
\value{

}
\description{
creat a new CellChat object
}
