% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{filterCommunication}
\alias{filterCommunication}
\title{Filter cell-cell communication if there are only few number of cells in certain cell groups}
\usage{
filterCommunication(object, min.cells = 10)
}
\arguments{
\item{object}{CellChat object}

\item{min.cells}{the minmum number of cells required in each cell group for cell-cell communication}
}
\value{
CellChat object with an updated slot net
}
\description{
Filter cell-cell communication if there are only few number of cells in certain cell groups
}
