% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{identifyOverExpressedGenes}
\alias{identifyOverExpressedGenes}
\title{Identify over-expressed genes in each cell group}
\usage{
identifyOverExpressedGenes(object, features = NULL, thresh.p = 0.05)
}
\arguments{
\item{object}{CellChat object}

\item{features}{a vector of features}

\item{thresh.p}{threshold of p-values}
}
\value{

}
\description{
Identify over-expressed genes in each cell group
}
