% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChat_class.R
\name{mergeCellChat}
\alias{mergeCellChat}
\title{Merge CellChat objects}
\usage{
mergeCellChat(object.list, add.names = NULL)
}
\arguments{
\item{object.list}{A list of multiple CellChat objects}

\item{add.names}{A vector containing the name of each dataset}
}
\value{

}
\description{
Merge CellChat objects
}
