% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{netAnalysis_computeCentrality}
\alias{netAnalysis_computeCentrality}
\title{Compute the network centrality scores allowing identification of dominant senders, receivers, mediators and influencers in all inferred communication networks}
\usage{
netAnalysis_computeCentrality(
  object = NULL,
  slot.name = "netP",
  net = NULL,
  net.name = NULL,
  thresh = 0.05
)
}
\arguments{
\item{object}{CellChat object; If object = NULL, USER must provide `net`}

\item{slot.name}{the slot name of object that is used to compute centrality measures of signaling networks}

\item{net}{compute the centrality measures on a specific signaling network given by a 2 or 3 dimemsional array net}

\item{net.name}{a character vector giving the name of signaling networks}

\item{thresh}{threshold of the p-value for determining significant interaction}
}
\value{

}
\description{
NB: This function was previously named as `netAnalysis_signalingRole`.  The previous function `netVisual_signalingRole` is now named as `netAnalysis_signalingRole_network`.
}
