% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{netAnalysis_signalingChanges_scatter}
\alias{netAnalysis_signalingChanges_scatter}
\title{2D visualization of differential outgoing and incoming signaling associated with one cell group}
\usage{
netAnalysis_signalingChanges_scatter(
  object,
  idents.use,
  color.use = c("grey10", "#F8766D", "#00BFC4"),
  comparison = c(1, 2),
  signaling = NULL,
  signaling.label = NULL,
  top.label = 1,
  signaling.exclude = NULL,
  slot.name = "netP",
  dot.size = 2.5,
  point.shape = c(21, 22, 24, 23),
  label.size = 3,
  dot.alpha = 0.6,
  xlabel = NULL,
  ylabel = NULL,
  title = NULL,
  font.size = 10,
  font.size.title = 10,
  do.label = T,
  show.legend = T,
  show.axes = T
)
}
\arguments{
\item{object}{A merged CellChat object}

\item{idents.use}{the cell group names of interest. Should be one of `levels(object@idents$joint)`}

\item{color.use}{a vector with three elements: the first is for coloring shared pathways, the second is for specific pathways in the first dataset, and the third is for specific pathways in the second dataset}

\item{comparison}{an index vector giving the two datasets for comparison}

\item{signaling}{a char vector containing signaling pathway names. signaling = NULL: Signaling role analysis on the aggregated cell-cell communication network from all signaling pathways}

\item{signaling.label}{a char vector giving the signaling names to show when labeling each point}

\item{top.label}{the fraction of signaling pathways to label}

\item{signaling.exclude}{signaling pathways to exclude when plotting}

\item{slot.name}{the slot name of object}

\item{dot.size}{a range defining the size of the symbol}

\item{point.shape}{point shape}

\item{label.size}{font size of the text}

\item{dot.alpha}{transparency}

\item{xlabel}{label of x-axis}

\item{ylabel}{label of y-axis}

\item{title}{main title of the plot}

\item{font.size}{font size of the text}

\item{font.size.title}{font size of the title}

\item{do.label}{label the each point}

\item{show.legend}{whether show the legend}

\item{show.axes}{whether show the axes}
}
\value{
ggplot object
}
\description{
2D visualization of differential outgoing and incoming signaling associated with one cell group
}
