% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{netEmbedding}
\alias{netEmbedding}
\title{Manifold learning of the signaling networks based on their similarity}
\usage{
netEmbedding(
  object,
  slot.name = "netP",
  type = c("functional", "structural"),
  pathway.remove = NULL,
  k = NULL
)
}
\arguments{
\item{object}{CellChat object}

\item{slot.name}{the slot name of object that is used to compute centrality measures of signaling networks}

\item{type}{"functional","structural"}

\item{pathway.remove}{a range of the number of patterns}

\item{k}{the number of nearest neighbors in running umap}
}
\value{

}
\description{
Manifold learning of the signaling networks based on their similarity
}
