% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual}
\alias{netVisual}
\title{Visualize the inferred cell-cell communication network}
\usage{
netVisual(
  object,
  signaling,
  signaling.name = NULL,
  color.use = NULL,
  vertex.receiver = NULL,
  sources.use = NULL,
  targets.use = NULL,
  top = 1,
  remove.isolate = FALSE,
  vertex.weight = 1,
  vertex.weight.max = NULL,
  vertex.size.max = NULL,
  weight.scale = TRUE,
  edge.weight.max.individual = NULL,
  edge.weight.max.aggregate = NULL,
  edge.width.max = 8,
  layout = c("circle", "hierarchy", "chord"),
  height = 5,
  thresh = 0.05,
  pt.title = 12,
  title.space = 6,
  vertex.label.cex = 0.8,
  from = NULL,
  to = NULL,
  bidirection = NULL,
  vertex.size = NULL,
  out.format = c("svg", "png"),
  group = NULL,
  cell.order = NULL,
  small.gap = 1,
  big.gap = 10,
  scale = FALSE,
  reduce = -1,
  show.legend = FALSE,
  legend.pos.x = 20,
  legend.pos.y = 20,
  nCol = NULL,
  ...
)
}
\arguments{
\item{object}{CellChat object}

\item{signaling}{a signaling pathway name}

\item{signaling.name}{alternative signaling pathway name to show on the plot}

\item{color.use}{the character vector defining the color of each cell group}

\item{vertex.receiver}{a numeric vector giving the index of the cell groups as targets in the first hierarchy plot}

\item{sources.use}{a vector giving the index or the name of source cell groups}

\item{targets.use}{a vector giving the index or the name of target cell groups.}

\item{top}{the fraction of interactions to show (0 < top <= 1)}

\item{remove.isolate}{whether remove the isolate nodes in the communication network}

\item{vertex.weight}{The weight of vertex: either a scale value or a vector

Default is a scale value being 1, indicating all vertex is plotted in the same size;

Set `vertex.weight` as a vector to plot vertex in different size; setting `vertex.weight = NULL` will have vertex with different size that are portional to the number of cells in each cell group.}

\item{vertex.weight.max}{the maximum weight of vertex; defualt = max(vertex.weight)}

\item{vertex.size.max}{the maximum vertex size for visualization}

\item{weight.scale}{whether scale the edge weight}

\item{edge.weight.max.individual}{the maximum weight of edge when plotting the individual L-R netwrok; defualt = max(net)}

\item{edge.weight.max.aggregate}{the maximum weight of edge when plotting the aggregated signaling pathway network}

\item{edge.width.max}{The maximum edge width for visualization}

\item{layout}{"hierarchy", "circle" or "chord"}

\item{height}{height of plot}

\item{thresh}{threshold of the p-value for determining significant interaction}

\item{pt.title}{font size of the text}

\item{title.space}{the space between the title and plot}

\item{vertex.label.cex}{The label size of vertex in the network}

\item{from, to, bidirection}{Deprecated. Use `sources.use`,`targets.use`}

\item{vertex.size}{Deprecated. Use `vertex.weight`

Parameters below are set for "chord" diagram. Please also check the function `netVisual_chord_cell` for more parameters.}

\item{out.format}{the format of output figures: svg, png and pdf}

\item{group}{A named group labels for making multiple-group Chord diagrams. The sector names should be used as the names in the vector.
The order of group controls the sector orders and if group is set as a factor, the order of levels controls the order of groups.}

\item{cell.order}{a char vector defining the cell type orders (sector orders)}

\item{small.gap}{Small gap between sectors.}

\item{big.gap}{Gap between the different sets of sectors, which are defined in the `group` parameter}

\item{scale}{scale each sector to same width; default = FALSE; however, it is set to be TRUE when remove.isolate = TRUE}

\item{reduce}{if the ratio of the width of certain grid compared to the whole circle is less than this value, the grid is removed on the plot. Set it to value less than zero if you want to keep all tiny grid.}

\item{show.legend}{whether show the figure legend}

\item{legend.pos.x, legend.pos.y}{adjust the legend position}

\item{nCol}{number of columns when displaying the network mediated by ligand-receptor using "circle" or "chord"}

\item{...}{other parameters (e.g.,vertex.label.cex, vertex.label.color, alpha.edge, label.edge, edge.label.color, edge.label.cex, edge.curved)
passing to `netVisual_hierarchy1`,`netVisual_hierarchy2`,`netVisual_circle`. NB: some parameters might be not supported}
}
\value{

}
\description{
Automatically save plots in the current working directory.
}
\examples{

}
