% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual_bubble}
\alias{netVisual_bubble}
\title{Show all the significant interactions (L-R pairs) from some cell groups to other cell groups}
\usage{
netVisual_bubble(
  object,
  from,
  to,
  pairLR.use0 = NULL,
  color.heatmap = viridis::viridis(50),
  thresh = 0.05
)
}
\arguments{
\item{object}{CellChat object}

\item{from}{a vector giving the index of source cell groups}

\item{to}{a vector giving the index of target cell groups}

\item{pairLR.use0}{ligand-receptor pairs to use; default is all the significant interactions}

\item{color.heatmap}{color map}

\item{thresh}{threshold of the p-value for determining significant interaction}
}
\value{

}
\description{
Show all the significant interactions (L-R pairs) from some cell groups to other cell groups
}
