% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{netVisual_heatmap}
\alias{netVisual_heatmap}
\title{Visualization of network using heatmap}
\usage{
netVisual_heatmap(
  object,
  comparison = c(1, 2),
  measure = c("count", "weight"),
  signaling = NULL,
  slot.name = c("netP", "net"),
  color.use = NULL,
  color.heatmap = c("#2166ac", "#b2182b"),
  title.name = NULL,
  legend.title = NULL,
  width = NULL,
  height = NULL,
  font.size = 8,
  font.size.title = 10,
  cluster.rows = FALSE,
  cluster.cols = FALSE,
  sources.use = NULL,
  targets.use = NULL,
  remove.isolate = FALSE,
  row.show = NULL,
  col.show = NULL
)
}
\arguments{
\item{object}{A merged CellChat object or a single CellChat object}

\item{comparison}{a numerical vector giving the datasets for comparison in object.list; e.g., comparison = c(1,2)}

\item{measure}{"count" or "weight". "count": comparing the number of interactions; "weight": comparing the total interaction weights (strength)}

\item{signaling}{a character vector giving the name of signaling networks in a single CellChat object}

\item{slot.name}{the slot name of object. Set is to be "netP" if input signaling is a pathway name; Set is to be "net" if input signaling is a ligand-receptor pair}

\item{color.use}{the character vector defining the color of each cell group}

\item{color.heatmap}{A vector of two colors corresponding to max/min values, or a color name in brewer.pal only when the data in the heatmap do not contain negative values}

\item{title.name}{the name of the title}

\item{legend.title}{the title of the heatmap legend}

\item{width}{width of the heatmap bidy, should be a fixed `unit` object}

\item{height}{height of the heatmap body, should be a fixed `unit` object}

\item{font.size}{fontsize in heatmap}

\item{font.size.title}{font size of the title}

\item{cluster.rows}{whether cluster rows}

\item{cluster.cols}{whether cluster columns}

\item{sources.use}{a vector giving the index or the name of source cell groups}

\item{targets.use}{a vector giving the index or the name of target cell groups.}

\item{remove.isolate}{whether remove the isolate nodes in the communication network}

\item{row.show, col.show}{a vector giving the index or the name of row or columns to show in the heatmap}
}
\value{
an object of ComplexHeatmap
}
\description{
This heatmap can be used to show differential number of interactions or interaction strength in the cell-cell communication network between two datasets;
the number of interactions or interaction strength in a single dataset
the inferred cell-cell communication network in single dataset, defined by `signaling`
}
\details{
When show differential number of interactions or interaction strength in the cell-cell communication network between two datasets, the width of edges represent the relative number of interactions or interaction strength.
Red (or blue) colored edges represent increased (or decreased) signaling in the second dataset compared to the first one.

The top colored bar plot represents the sum of column of values displayed in the heatmap. The right colored bar plot represents the sum of row of values.
}
