% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize data using a scaling factor}
\usage{
normalizeData(data.raw, scale.factor = 10000, do.log = TRUE, do.sparse = TRUE)
}
\arguments{
\item{data.raw}{input raw data}

\item{scale.factor}{the scaling factor used for each cell}

\item{do.log}{whether to do log transformation with pseudocount 1}

\item{do.sparse}{whether to use sparse format}
}
\description{
Normalize data using a scaling factor
}
