% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{rankNet}
\alias{rankNet}
\title{Rank signaling networks based on the information flow}
\usage{
rankNet(
  object,
  slot.name = "netP",
  mode = c("single", "comparison"),
  comparison = c(1, 2),
  tol = 0.05,
  return.data = FALSE,
  show.raw = FALSE,
  x.rotation = 90,
  title = NULL,
  color.use = NULL,
  bar.w = 0.75,
  font.size = 8
)
}
\arguments{
\item{object}{CellChat object}

\item{slot.name}{the slot name of object that is used to compute centrality measures of signaling networks}

\item{mode}{"single","comparison"}

\item{comparison}{a numerical vector giving the datasets for comparison; a single value means ranking for only one dataset and two values means ranking comparison for two datasets}

\item{tol}{a tolerance when considering the relative contribution being equal between two datasets. contribution.relative between 1-tol and 1+tol will be considered as equal contribution}

\item{return.data}{whether return the data.frame consisting of the calculated information flow of each signaling pathway or L-R pair}

\item{show.raw}{whether show the raw information flow. Default = FALSE, showing the scaled information flow to provide compariable data scale}

\item{x.rotation}{rotation of x-labels}

\item{title}{main title of the plot}

\item{color.use}{defining the color for each cell group}

\item{bar.w}{the width of bar plot}

\item{font.size}{font size}
}
\value{

}
\description{
Rank signaling networks based on the information flow
}
