% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{sketchData}
\alias{sketchData}
\title{Downsampling single cell data using geometric sketching algorithm}
\usage{
sketchData(X, percent, dimPC = 50)
}
\arguments{
\item{X}{input data (samples in rows, features in columns)}

\item{percent}{the percent of data to sketch}

\item{dimPC}{the number of components to use}
}
\value{
A vector consisting of the index to use for downsampling
}
\description{
USERs need to install the python package `pip install geosketch` (https://github.com/brianhie/geosketch)
}
