% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{subsetCommunication}
\alias{subsetCommunication}
\title{Subset the inferred cell-cell communications of interest}
\usage{
subsetCommunication(
  object,
  slot.name = "net",
  sources.use = NULL,
  targets.use = NULL,
  signaling = NULL,
  pairLR.use = NULL,
  thresh = 0.05
)
}
\arguments{
\item{object}{CellChat object}

\item{slot.name}{the slot name of object: slot.name = "net" when extracting the inferred communications at the level of ligands/receptors; slot.name = "netP" when extracting the inferred communications at the level of signaling pathways}

\item{sources.use}{a vector giving the index or the name of source cell groups}

\item{targets.use}{a vector giving the index or the name of target cell groups.}

\item{signaling}{a character vector giving the name of signaling pathways of interest}

\item{pairLR.use}{a data frame consisting of one column named either "interaction_name" or "pathway_name", defining the interactions of interest}

\item{thresh}{threshold of the p-value for determining significant interaction}
}
\value{
If input object is created from a single dataset, a data frame of the inferred cell-cell communications of interest, consisting of source, target, interaction_name, pathway_name, prob and other information

If input object is a merged object from multiple datasets, it will return a list and each element is a data frame for one dataset
}
\description{
NB: If all arguments are NULL, it returns a data frame consisting of all the inferred cell-cell communications
}
\examples{
\dontrun{
# access all the inferred cell-cell communications
df.net <- subsetCommunication(cellchat)

# access all the inferred cell-cell communications at the level of signaling pathways
df.net <- subsetCommunication(cellchat, slot.name = "netP")

# Subset to certain cells with sources.use and targets.use
df.net <- subsetCommunication(cellchat, sources.use = c(1,2), targets.use = c(4,5))

# Subset to certain signaling, e.g., WNT and TGFb
df.net <- subsetCommunication(cellchat, signaling = c("WNT", "TGFb"))
}

}
