% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{subsetCommunication_internal}
\alias{subsetCommunication_internal}
\title{Subset the inferred cell-cell communications of interest}
\usage{
subsetCommunication_internal(
  net,
  LR,
  cells.level,
  slot.name = "net",
  sources.use = NULL,
  targets.use = NULL,
  signaling = NULL,
  pairLR.use = NULL,
  thresh = 0.05
)
}
\arguments{
\item{net, LR, cells.level}{object@net object@LR levels(object@idents)}

\item{slot.name}{the slot name of object: slot.name = "net" when extracting the inferred communications at the level of ligands/receptors; slot.name = "netP" when extracting the inferred communications at the level of signaling pathways}

\item{sources.use}{a vector giving the index or the name of source cell groups}

\item{targets.use}{a vector giving the index or the name of target cell groups.}

\item{signaling}{a character vector giving the name of signaling pathways of interest}

\item{pairLR.use}{a data frame consisting of one column named either "interaction_name" or "pathway_name", defining the interactions of interest}

\item{thresh}{threshold of the p-value for determining significant interaction}
}
\value{
A data frame of the inferred cell-cell communications of interest, consisting of source, target, interaction_name, pathway_name, prob and other information
}
\description{
NB: If all arguments are NULL, it returns a data frame consisting of all the inferred cell-cell communications
}
