% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{subsetDB}
\alias{subsetDB}
\title{Subset CellChatDB databse by only including interactions of interest}
\usage{
subsetDB(CellChatDB, search = c(), key = "annotation")
}
\arguments{
\item{CellChatDB}{CellChatDB databse}

\item{search}{a character}

\item{key}{the name of the variable in CellChatDB interaction_input}
}
\value{

}
\description{
Subset CellChatDB databse by only including interactions of interest
}
