% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellChat_class.R
\name{updateCellChat}
\alias{updateCellChat}
\title{Update a single CellChat object}
\usage{
updateCellChat(object)
}
\arguments{
\item{object}{CellChat object}
}
\value{
a updated CellChat object
}
\description{
Update a single previously calculated CellChat object for spatial transcriptomics data analysis (version < 2.1.0)
}
\details{
Update a single previously calculated CellChat object (version < 1.6.0)

version < 0.5.0: `object@var.features` is now `object@var.features$features`; `object@net$sum` is now `object@net$weight` if `aggregateNet` has been run.

version 1.6.0: a `object@images` slot is added and `datatype` is added in `object@options$datatype`

version 2.1.0: a column named `slices` is added in `meta` data for spatial transcriptomics data analysis.

version 2.1.1: `images$scale.factors` is changed to `images$spatial.factors` for spatial transcriptomics data analysis.

version 2.1.2: the column `slices` in `object@meta` is renamed as `samples` in order to identify consistent signaling across samples for cell-cell communication analysis.

version 2.1.3: the slot `object@data.project` is renamed as `object@data.smooth`.
}
